feat: 실제 users 엔티티에 맞춰 data, schema.sql 및 테스트코드 수정
